upd project structure, refactor old files, fix license mail
