+++ Src/dronecan_application/modules/CircuitStatus/CMakeLists.txt
@@ -1,0 +1,8 @@
+set(CIRCUITSTATUS_MODULE_DIR ${CMAKE_CURRENT_LIST_DIR})
+
+set(CIRCUITSTATUS_HEADERS 
+    ${CIRCUITSTATUS_MODULE_DIR}
+)
+set(CIRCUITSTATUS_SOURCES
+    ${CIRCUITSTATUS_MODULE_DIR}/CircuitStatus.cpp
+)
