+++ Src/dronecan_application/modules/CircuitStatus/CircuitStatus.hpp
@@ -1,20 +1,19 @@
 /***
- * Copyright (C) 2024 Anastasiia Stepanova  <asiiapine96@gmail.com>
- *  Distributed under the terms of the GPL v3 license, available in the file LICENSE.
-***/ 
+ * Copyright (C) 2024 Anastasiia Stepanova  <asiiapine@gmail.com>
+ *  Distributed under the terms of the GPL v3 license, available in the file
+ *LICENSE.
+ ***/
 
 #ifndef SRC_MODULE_CIRCUIT_STATUS_HPP_
 #define SRC_MODULE_CIRCUIT_STATUS_HPP_
 
-#include "periphery/led/led.hpp"
+#include "logger.hpp"
+#include "periphery/adc/adc.hpp"
 #include "uavcan/equipment/power/CircuitStatus.h"
 #include "uavcan/equipment/temperature/Temperature.h"
-#include "periphery/adc/adc.hpp"
-#include "logger.hpp"
-
 
 class CircuitStatusModule {
-public:
+   public:
     static CircuitStatusModule &get_instance();
     void spin_once();
 
@@ -23,31 +22,31 @@ public:
     float volt = 0;
     float curr = 0;
 
-private:
+   private:
     static CircuitStatusModule instance;
     static bool instance_initialized;
 
     CircuitStatus_t circuit_status = {};
     Temperature_t temperature_status = {};
-    
+
     static Logger logger;
 
-    CircuitStatusModule(){};
+    CircuitStatusModule(){}
     AdcPeriphery adc;
 
-    uint8_t circuit_status_transfer_id  = 0;
-    uint8_t temperature_transfer_id     = 0;
+    uint8_t circuit_status_transfer_id = 0;
+    uint8_t temperature_transfer_id = 0;
 
-    uint16_t v5_raw     = 0;
-    uint16_t cur_raw    = 0;
-    uint16_t vol_raw    = 0;
-    uint16_t tem_raw    = 0;
+    uint16_t v5_raw = 0;
+    uint16_t cur_raw = 0;
+    uint16_t vol_raw = 0;
+    uint16_t tem_raw = 0;
 
     int8_t publish_error = 0;
 
     void init();
-    CircuitStatusModule(CircuitStatusModule &other) = delete;
+    CircuitStatusModule(const CircuitStatusModule &other) = delete;
     void operator=(const CircuitStatusModule &) = delete;
 };
 
-#endif //SRC_MODULE_CIRCUIT_STATUS_HPP_
+#endif  // SRC_MODULE_CIRCUIT_STATUS_HPP_
