+++ Src/platform/stm32f103/pwm.cpp
@@ -1,5 +1,5 @@
 /***
- * Copyright (C) 2024 Anastasiia Stepanova  <asiiapine96@gmail.com>
+ * Copyright (C) 2024 Anastasiia Stepanova  <asiiapinemail.com>
  *  Distributed under the terms of the GPL v3 license, available in the file
  *LICENSE.
  ***/
